refactor: modularize MCP tool registration
