+++ pyproject.toml
@@ -49,17 +49,11 @@ include = [
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/renforge"]
-exclude = ["src/renforge/ui/static/**"]
-
-[tool.hatch.build.targets.wheel.force-include]
-"src/renforge/ui/static" = "renforge/ui/static"
+artifacts = ["src/renforge/ui/static/**"]
 
 [tool.hatch.build.targets.sdist]
 include = ["src/renforge"]
-exclude = ["src/renforge/ui/static/**"]
-
-[tool.hatch.build.targets.sdist.force-include]
-"src/renforge/ui/static" = "src/renforge/ui/static"
+artifacts = ["src/renforge/ui/static/**"]
 
 [tool.pytest.ini_options]
 pythonpath = ["src"]
