+++ src/renforge/server.py
@@ -2,21 +2,9 @@
 
 from __future__ import annotations
 
-import base64
-import hashlib
-import math
-import threading
 from dataclasses import dataclass
-from inspect import signature
 from pathlib import Path
-from time import perf_counter
-from typing import Annotated, Any, Callable, Literal, Optional, get_args, get_origin, get_type_hints
-
-from pydantic import Field
-
-from . import __version__, session_registry
-from .project import discover_project_from
-from .tool_definitions import TOOL_DEFINITIONS
+from typing import Any, Optional
 
 
 @dataclass
@@ -49,1835 +37,10 @@ def _get_fastmcp_backend() -> tuple[Optional[type], Optional[str]]:
         return None, None
 
 
-def _log_tool_call(
-    *,
-    name: str,
-    params: dict[str, Any],
-    project_root: str | None,
-    fn: Callable[..., Any],
-    args: tuple[Any, ...],
-    kwargs: dict[str, Any],
-) -> Any:
-    from . import activity_log
-
-    started = perf_counter()
-    try:
-        result = fn(*args, **kwargs)
-    except Exception as exc:
-        result = {"ok": False, "error": f"{type(exc).__name__}: {exc}"}
-
-    duration_ms = round((perf_counter() - started) * 1000, 2)
-    definition = TOOL_DEFINITIONS.get(name)
-    should_log = definition is None or not definition.annotations["readOnlyHint"]
-    if project_root is not None and should_log:
-        summary = activity_log.summarize_result(result)
-        try:
-            activity_log.log_tool_call(
-                project_root,
-                name,
-                params,
-                duration_ms,
-                result,
-                files_touched=summary["files_touched"],
-            )
-        except Exception:
-            pass
-
-    return result
-
-
-def _png_content(png: bytes) -> Any:
-    from mcp.types import ImageContent
-
-    return ImageContent(
-        type="image",
-        data=base64.b64encode(png).decode("ascii"),
-        mimeType="image/png",
-    )
-
-
 def _register_tools(app: Any) -> None:
-    tool_decorator: Callable[..., Any] | None = getattr(app, "tool", None)
-    if not callable(tool_decorator):
-        return
+    from .tool_registration import register_all_tools
 
-    from .tools import live, project_ops
-    from .tools.static import inspect_project, parse_lint_text, scan_project_index
-
-    try:
-        from mcp.types import ToolAnnotations
-    except Exception:  # pragma: no cover - compatibility with older backend module layouts
-        ToolAnnotations = None
-
-    registered_tools: dict[str, bool] = {}
-
-    def _to_annotations(raw: dict[str, bool]) -> Any:
-        if ToolAnnotations is None:
-            return raw
-        return ToolAnnotations(
-            readOnlyHint=raw["readOnlyHint"],
-            idempotentHint=raw["idempotentHint"],
-            destructiveHint=raw["destructiveHint"],
-            openWorldHint=raw["openWorldHint"],
-        )
-
-    def _annotate_parameters(
-        fn: Callable[..., Any],
-        metadata: Any,
-    ) -> Callable[..., Any]:
-        resolved_hints = get_type_hints(fn, include_extras=True)
-        decorated_annotations: dict[str, Any] = {}
-        for param, annotation in resolved_hints.items():
-            if param == "return":
-                continue
-            annotation = metadata.parameter_types.get(param, annotation)
-            schema = metadata.parameter_schemas.get(param, {})
-            enum_values = schema.get("enum")
-            if enum_values:
-                literal = Literal.__getitem__(tuple(enum_values))
-                if type(None) in get_args(annotation):
-                    annotation = literal | None
-                else:
-                    annotation = literal
-            item_enum = (schema.get("items") or {}).get("enum")
-            if item_enum:
-                literal_item = Literal.__getitem__(tuple(item_enum))
-                origin = get_origin(annotation)
-                args = get_args(annotation)
-                if origin is list:
-                    annotation = list[literal_item]
-                elif type(None) in args:
-                    annotation = list[literal_item] | None
-            decorated_annotations[param] = Annotated[
-                annotation,
-                Field(
-                    description=metadata.parameters[param],
-                    ge=schema.get("minimum"),
-                    le=schema.get("maximum"),
-                    pattern=schema.get("pattern"),
-                    json_schema_extra=schema or None,
-                ),
-            ]
-        if "return" in resolved_hints:
-            decorated_annotations["return"] = resolved_hints["return"]
-        fn.__annotations__ = decorated_annotations
-        return fn
-
-    def _register_tool(fn: Callable[..., Any]) -> Callable[..., Any]:
-        name = fn.__name__
-        metadata = TOOL_DEFINITIONS.get(name)
-        if metadata is None:
-            raise RuntimeError(f"Missing tool definition for {name}")
-
-        sig = signature(fn)
-        expected_parameters = set(sig.parameters)
-        documented_parameters = set(metadata.parameters)
-        if expected_parameters != documented_parameters:
-            missing = sorted(expected_parameters - documented_parameters)
-            extra = sorted(documented_parameters - expected_parameters)
-            raise RuntimeError(
-                f"Tool definition drift for {name}: params mismatch "
-                f"missing={missing} extra={extra}"
-            )
-
-        annotated_tool = _annotate_parameters(fn, metadata)
-        annotated_tool.__doc__ = metadata.description
-
-        registered_tools[name] = True
-        decorator_kwargs = {
-            "description": metadata.description,
-            "annotations": _to_annotations(metadata.annotations),
-        }
-        try:
-            decorator_signature = signature(tool_decorator)
-        except (TypeError, ValueError):
-            if not getattr(app, "_renforge_testing_registry", False):
-                raise RuntimeError(
-                    "MCP tool backend cannot prove support for required description and annotations metadata"
-                )
-            supported_kwargs = {}
-        else:
-            accepts_arbitrary_kwargs = any(
-                p.kind == p.VAR_KEYWORD for p in decorator_signature.parameters.values()
-            )
-            if accepts_arbitrary_kwargs:
-                supported_kwargs = decorator_kwargs
-            else:
-                supported_kwargs = {
-                    key: value
-                    for key, value in decorator_kwargs.items()
-                    if key in decorator_signature.parameters
-                }
-            missing_metadata = set(decorator_kwargs) - set(supported_kwargs)
-            if missing_metadata and not getattr(app, "_renforge_testing_registry", False):
-                raise RuntimeError(
-                    "MCP tool backend cannot accept required description and annotations metadata"
-                )
-        registered = tool_decorator(**supported_kwargs)(annotated_tool)
-        component = None
-        local_provider = getattr(app, "_local_provider", None)
-        components = getattr(local_provider, "_components", None)
-        if isinstance(components, dict):
-            component = next(
-                (item for item in components.values() if getattr(item, "name", None) == name),
-                None,
-            )
-        tool_manager = getattr(app, "_tool_manager", None)
-        managed_tools = getattr(tool_manager, "_tools", None)
-        if component is None and isinstance(managed_tools, dict):
-            component = managed_tools.get(name)
-        parameters = getattr(component, "parameters", None)
-        if isinstance(parameters, dict):
-            for param, schema in metadata.parameter_schemas.items():
-                if schema.get("type") != "object":
-                    continue
-                property_schema = parameters.get("properties", {}).get(param)
-                if not isinstance(property_schema, dict) or "anyOf" not in property_schema:
-                    continue
-                description = property_schema.get("description")
-                default = property_schema.get("default")
-                property_schema.clear()
-                property_schema["anyOf"] = [schema, {"type": "null"}]
-                property_schema["default"] = default
-                property_schema["description"] = description
-            parameters.update(metadata.input_schema)
-        return registered
-
-    def _launch_game(
-        project_path: str,
-        *,
-        version: str = "stable",
-        warp: str | None = None,
-        editor: bool = True,
-        display: str = "auto",
-        audio: str = "auto",
-        savedir: str | None = None,
-        persistent: str = "existing",
-        cleanup_on_stop: bool = True,
-        timeout: float | None = None,
-        session: dict[str, Any] | None = None,
-        cancel_event: threading.Event | None = None,
-    ) -> dict:
-        canonical_editor = bool(editor)
-        session_cfg = dict(session or {})
-        effective_savedir = session_cfg.get("savedir", savedir)
-        effective_persistent = str(session_cfg.get("persistent", persistent) or "existing")
-        effective_cleanup = (
-            session_cfg["cleanup_on_stop"]
-            if isinstance(session_cfg.get("cleanup_on_stop"), bool)
-            else cleanup_on_stop
-        )
-        if cancel_event is not None and cancel_event.is_set():
-            return live.cancelled_launch_result(phase="detecting_environment")
-        from .dashboard_client import (
-            launch_game as launch_via_dashboard,
-            stop_game as stop_via_dashboard,
-        )
-
-        # None = no matching dashboard. Any dict (including failure) is final:
-        # never fall back to a local launch after a contacted dashboard errors.
-        delegated = launch_via_dashboard(
-            project_path,
-            version=version,
-            warp=warp,
-            editor=canonical_editor,
-            display=display,
-            audio=audio,
-            savedir=effective_savedir if isinstance(effective_savedir, str) else None,
-            persistent=effective_persistent,
-            cleanup_on_stop=bool(effective_cleanup),
-            timeout=timeout,
-        )
-        if delegated is not None:
-            if cancel_event is not None and cancel_event.is_set():
-                stopped = stop_via_dashboard(project_path)
-                if stopped is None:
-                    message = (
-                        "The dashboard launch was cancelled, but its owning "
-                        "dashboard is unavailable to stop it."
-                    )
-                    return {
-                        "ok": False,
-                        "ready": False,
-                        "code": "DASHBOARD_STOP_UNAVAILABLE",
-                        "phase": "stopping_cancelled_dashboard_launch",
-                        "message": message,
-                        "error": message,
-                        "launch_cancel_requested": True,
-                    }
-                if not stopped.get("ok"):
-                    return {
-                        **stopped,
-                        "ready": False,
-                        "launch_cancel_requested": True,
-                    }
-                return live.cancelled_launch_result(phase="starting_renpy")
-            return delegated
-        if cancel_event is not None and cancel_event.is_set():
-            stopped = live.stop_external_game(project_path)
-            if not stopped.get("ok"):
-                return {
-                    **stopped,
-                    "ready": False,
-                    "launch_cancel_requested": True,
-                }
-            return live.cancelled_launch_result(phase="starting_renpy")
-        return live.launch_game(
-            project_path,
-            version=version,
-            warp=warp,
-            editor=canonical_editor,
-            display=display,
-            audio=audio,
-            savedir=effective_savedir if isinstance(effective_savedir, str) else None,
-            persistent=effective_persistent,
-            cleanup_on_stop=bool(effective_cleanup),
-            timeout=timeout,
-            session=session,
-            cancel_event=cancel_event,
-        )
-
-    def _stop_game(project_path: str) -> dict:
-        from .dashboard_client import stop_game as stop_via_dashboard
-
-        delegated = stop_via_dashboard(project_path)
-        # Contacted dashboard failure is final; None means no dashboard.
-        return delegated if delegated is not None else live.stop_game(project_path)
-
-    def _launch_status(project_path: str) -> dict:
-        from .dashboard_client import launch_status as status_via_dashboard
-
-        delegated = status_via_dashboard(project_path)
-        return delegated if delegated is not None else live.launch_status(project_path)
-
-    def _start_launch(project_path: str, **kwargs: Any) -> dict:
-        requested_editor = bool(kwargs.get("editor", True))
-        kwargs["editor"] = requested_editor
-
-        def _launch(project_root: Path, cancel_event: threading.Event) -> dict:
-            return _launch_game(
-                str(project_root),
-                cancel_event=cancel_event,
-                **kwargs,
-            )
-
-        return live.start_launch(project_path, _launch, editor=requested_editor)
-
-    def _context_payload() -> dict[str, Any]:
-        dashboard = session_registry.active_dashboard()
-        default_project = getattr(app, "project_root", None)
-        active_project = dashboard.get("project") if dashboard else None
-        project_source = "dashboard" if active_project else None
-        if active_project is None and default_project is not None:
-            active_project = str(Path(default_project).expanduser().resolve())
-            project_source = "serve_default"
-        if active_project is None:
-            detected = discover_project_from()
-            if detected is not None:
-                active_project = str(detected)
-                project_source = "cwd"
-        payload: dict[str, Any] = {
-            "ok": True,
-            "version": __version__,
-            "active_project": active_project,
-            "project_source": project_source,
-            "dashboard": dashboard,
-            "live_editor": {
-                "enabled_by_default": True,
-                "launch_tool": "renforge_launch",
-                "guide": "docs/LIVE_EDITOR.md",
-                "summary": (
-                    "In-game Live Editor is injected by default on "
-                    "renforge_launch. Preview is runtime-only until Save; "
-                    "locked targets stay inspectable. Use only public MCP "
-                    "tools — never private editor_task0_* handlers."
-                ),
-                "agent_workflow": [
-                    "renforge_info",
-                    "renforge_launch",
-                    "renforge_launch_status",
-                    "renforge_screenshot",
-                    "renforge_scene_tree",
-                    "renforge_click_at",
-                    "renforge_click_element",
-                    "renforge_stop",
-                ],
-            },
-        }
-        if active_project is None:
-            payload["hint"] = (
-                "No dashboard, serve default, or Ren'Py project near the "
-                "current directory. Every tool accepts project_path directly: "
-                "ask the user for the game's path."
-            )
-        return payload
-
-    @_register_tool
-    def renforge_info() -> dict:
-        """Call first: report RenForge version and the active project.
-
-        active_project falls back from the dashboard selection to the serve
-        default, then to a Ren'Py project detected from the current directory
-        (project_source says which one matched). A null active_project only
-        means auto-discovery found nothing — every tool accepts project_path
-        directly, so ask the user for the game's path and keep going.
-        """
-        return _context_payload()
-
-    @_register_tool
-    def renforge_context() -> dict:
-        """Discover the active Ren'Py project (dashboard, serve default, or cwd)."""
-        return _context_payload()
-
-    @_register_tool
-    def renforge_inspect_image(
-        image_path: str,
-        crop_x: int = 0,
-        crop_y: int = 0,
-        crop_width: int = 0,
-        crop_height: int = 0,
-        scale: float = 1.0,
-    ):
-        """Open a local image and return an optional cropped/zoomed PNG for inspection."""
-        from .image_ops import inspect_image
-
-        return _log_tool_call(
-            name="renforge_inspect_image",
-            params={
-                "image_path": image_path,
-                "crop_x": crop_x,
-                "crop_y": crop_y,
-                "crop_width": crop_width,
-                "crop_height": crop_height,
-                "scale": scale,
-            },
-            project_root=None,
-            fn=lambda: _png_content(
-                inspect_image(
-                    image_path,
-                    crop_x=crop_x,
-                    crop_y=crop_y,
-                    crop_width=crop_width,
-                    crop_height=crop_height,
-                    scale=scale,
-                )
-            ),
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_inspect_project(project_path: str) -> dict:
-        return _log_tool_call(
-            name="renforge_inspect_project",
-            params={"project_path": project_path},
-            project_root=project_path,
-            fn=inspect_project,
-            args=(project_path,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_scan_project(
-        project_path: str,
-        sections: list[str] | None = None,
-        file_glob: str = "",
-        symbol: str = "",
-        offset: int = 0,
-        limit: int = 200,
-    ) -> dict:
-        """Scan scripts; defaults to summary-only, with opt-in sections and pagination."""
-        selected_sections = [] if sections is None else sections
-        return _log_tool_call(
-            name="renforge_scan_project",
-            params={
-                "project_path": project_path,
-                "sections": selected_sections,
-                "file_glob": file_glob,
-                "symbol": symbol,
-                "offset": offset,
-                "limit": limit,
-            },
-            project_root=project_path,
-            fn=scan_project_index,
-            args=(project_path,),
-            kwargs={
-                "sections": selected_sections,
-                "file_glob": file_glob,
-                "symbol": symbol,
-                "offset": offset,
-                "limit": limit,
-            },
-        )
-
-    @_register_tool
-    def renforge_find_references(
-        project_path: str,
-        symbol: str,
-        file_glob: str = "",
-        offset: int = 0,
-        limit: int = 200,
-    ) -> dict:
-        """Find exact Ren'Py definitions/usages, including text interpolations."""
-        from .symbols import find_references
-
-        return _log_tool_call(
-            name="renforge_find_references",
-            params={
-                "project_path": project_path,
-                "symbol": symbol,
-                "file_glob": file_glob,
-                "offset": offset,
-                "limit": limit,
-            },
-            project_root=project_path,
-            fn=find_references,
-            args=(project_path, symbol),
-            kwargs={"file_glob": file_glob, "offset": offset, "limit": limit},
-        )
-
-    @_register_tool
-    def renforge_parse_lint(text: str) -> dict:
-        return _log_tool_call(
-            name="renforge_parse_lint",
-            params={"text": text},
-            project_root=None,
-            fn=parse_lint_text,
-            args=(text,),
-            kwargs={},
-        )
-
-    # --- live game control (requires a display; under WSLg it works directly) ---
-
-    @_register_tool
-    def renforge_launch(
-        project_path: str,
-        warp: str = "",
-        version: str = "stable",
-        editor: bool = True,
-        display: str = "auto",
-        audio: str = "auto",
-        savedir: str = "",
-        persistent: str = "existing",
-        cleanup_on_stop: bool = True,
-        timeout: float = 0,
-    ) -> dict:
-        """Launch or reuse a game with the Live Editor enabled by default.
-
-        Pass ``editor=False`` to launch intentionally without the visual editor.
-
-        After launch, poll ``renforge_launch_status`` until ready, then observe
-        with a fresh ``renforge_screenshot`` or ``renforge_scene_tree`` before
-        any click. Use ``renforge_click_at`` or ``renforge_click_element`` with
-        a current ``frame_id`` guard; verify the result, then ``renforge_stop``.
-        See docs/LIVE_EDITOR.md. Optional ``warp`` is a Ren'Py file:line target.
-
-        The call waits at most 20 seconds for readiness, then returns
-        ``status="starting"`` while startup continues in the background. Poll
-        ``renforge_launch_status`` until it reports ``ready`` or ``failed``.
-        display/audio default to auto; savedir='temporary' isolates saves.
-        timeout controls the background startup deadline, not the MCP call.
-        """
-        kwargs: dict[str, Any] = {
-            "version": version,
-            "warp": warp or None,
-            "editor": editor,
-            "display": display or "auto",
-            "audio": audio or "auto",
-            "persistent": persistent or "existing",
-            "cleanup_on_stop": cleanup_on_stop,
-        }
-        if savedir:
-            kwargs["savedir"] = savedir
-        if timeout and timeout > 0:
-            kwargs["timeout"] = float(timeout)
-        return _log_tool_call(
-            name="renforge_launch",
-            params={
-                "project_path": project_path,
-                "warp": warp,
-                "version": version,
-                "editor": editor,
-                "display": display,
-                "audio": audio,
-                "savedir": savedir,
-                "persistent": persistent,
-                "cleanup_on_stop": cleanup_on_stop,
-                "timeout": timeout,
-            },
-            project_root=project_path,
-            fn=_start_launch,
-            args=(project_path,),
-            kwargs=kwargs,
-        )
-
-    @_register_tool
-    def renforge_launch_status(project_path: str) -> dict:
-        """Return starting, ready, failed, or idle for a background launch."""
-        return _log_tool_call(
-            name="renforge_launch_status",
-            params={"project_path": project_path},
-            project_root=project_path,
-            fn=_launch_status,
-            args=(project_path,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_jump(project_path: str, target: str, version: str = "stable") -> dict:
-        """Restart at a label or file:line; poll launch status when still starting."""
-        from .navigation import resolve_warp_target
-
-        def _jump() -> dict:
-            resolved = resolve_warp_target(project_path, target)
-            if not resolved.get("ok"):
-                return resolved
-            return _start_launch(
-                project_path,
-                version=version,
-                warp=str(resolved["target"]),
-            )
-
-        return _log_tool_call(
-            name="renforge_jump",
-            params={"project_path": project_path, "target": target, "version": version},
-            project_root=project_path,
-            fn=_jump,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_new_game(project_path: str, version: str = "stable") -> dict:
-        """Start at the ``start`` label; poll launch status when still starting."""
-        from .navigation import resolve_warp_target
-
-        def _new_game() -> dict:
-            resolved = resolve_warp_target(project_path, "start")
-            if not resolved.get("ok"):
-                return resolved
-            return _start_launch(
-                project_path,
-                version=version,
-                warp=str(resolved["target"]),
-            )
-
-        return _log_tool_call(
-            name="renforge_new_game",
-            params={"project_path": project_path, "version": version},
-            project_root=project_path,
-            fn=_new_game,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_stop(project_path: str) -> dict:
-        """Stop a running game or cancel its in-progress launch, then clean up."""
-        return _log_tool_call(
-            name="renforge_stop",
-            params={"project_path": project_path},
-            project_root=project_path,
-            fn=_stop_game,
-            args=(project_path,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_game_state(
-        project_path: str,
-        include: list[str] | None = None,
-    ) -> dict:
-        """Return complete live state; optionally include compact metrics or audio."""
-        def _state() -> dict:
-            # Preserve the no-payload wire shape for existing callers.
-            if include is None:
-                return live.game_state(project_path)
-            return live.game_state(project_path, include=include)
-
-        return _log_tool_call(
-            name="renforge_game_state",
-            params={"project_path": project_path, "include": include},
-            project_root=project_path,
-            fn=_state,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_game_state_compact(
-        project_path: str,
-        variable_names: list[str] | None = None,
-        variable_prefix: str = "",
-        state_profile: str = "interaction",
-        max_depth: int = 3,
-        max_items: int = 50,
-        max_output_bytes: int = 8192,
-    ) -> dict:
-        """Return bounded live state, optionally with selected variables.
-
-        Defaults to state_profile='interaction' so the full store is never
-        returned in the payload unless state_profile='full' is requested.
-        """
-        def _state() -> dict:
-            from .state_compact import (
-                apply_serialization_limits,
-                compact_state,
-                normalize_state_profile,
-                validate_limit_args,
-            )
-
-            state = live.game_state(project_path)
-            if not state.get("ok"):
-                return state
-            profile = normalize_state_profile(state_profile, default="interaction")
-            if isinstance(profile, dict):
-                return profile
-            limits = validate_limit_args(
-                max_depth=max_depth,
-                max_items=max_items,
-                max_output_bytes=max_output_bytes,
-            )
-            if isinstance(limits, dict):
-                return limits
-            depth, items, budget = limits
-
-            variables = state.get("variables", {})
-            if not isinstance(variables, dict):
-                variables = {}
-            variable_count = len(variables)
-            requested = set(variable_names or [])
-            if requested or variable_prefix:
-                variables = {
-                    name: value
-                    for name, value in variables.items()
-                    if name in requested or (variable_prefix and name.startswith(variable_prefix))
-                }
-            # Rebuild a state object for compact_state with the filtered vars.
-            source = dict(state)
-            source["variables"] = variables
-            include = list(variables) if (requested or variable_prefix) else list(variable_names or [])
-            compacted = compact_state(
-                source,
-                profile=profile if profile != "full" else "full",
-                include=include or None,
-                max_depth=depth,
-                max_items=items,
-                max_output_bytes=budget,
-            )
-            result = {"ok": True, "state_profile": profile, "variable_count": variable_count}
-            result.update(compacted)
-            result["variable_count"] = variable_count
-            limited_result = apply_serialization_limits(
-                result,
-                max_depth=depth,
-                max_items=items,
-                max_output_bytes=budget,
-            )
-            return limited_result if isinstance(limited_result, dict) else {"value": limited_result}
-
-        return _log_tool_call(
-            name="renforge_game_state_compact",
-            params={
-                "project_path": project_path,
-                "variable_names": variable_names,
-                "variable_prefix": variable_prefix,
-                "state_profile": state_profile,
-                "max_depth": max_depth,
-                "max_items": max_items,
-                "max_output_bytes": max_output_bytes,
-            },
-            project_root=project_path,
-            fn=_state,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_inspect_screen(project_path: str, name: str) -> dict:
-        """Inspect an active screen's layer, JSON-safe scope, and arguments."""
-        return _log_tool_call(
-            name="renforge_inspect_screen",
-            params={"project_path": project_path, "name": name},
-            project_root=project_path,
-            fn=live.inspect_screen,
-            args=(project_path, name),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_advance(project_path: str) -> dict:
-        """Advance the current dialogue."""
-        return _log_tool_call(
-            name="renforge_advance",
-            params={"project_path": project_path},
-            project_root=project_path,
-            fn=live.advance,
-            args=(project_path,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_control(
-        project_path: str,
-        action: str,
-        interaction_id: str = "",
-        wait_for_effect: bool = False,
-        effect_timeout: float = 5.0,
-    ) -> dict:
-        """Run a runtime action: advance, rollback, toggle_skip, toggle_auto,
-        toggle_afm, game_menu, hide_windows, quick_save, quick_load,
-        reload_script, restart_interaction, or quit.
-
-        Emits correlated business events (quick_save.completed, skip.stopped,
-        …). Set wait_for_effect=true to block until the matching event appears.
-        """
-        return _log_tool_call(
-            name="renforge_control",
-            params={
-                "project_path": project_path,
-                "action": action,
-                "interaction_id": interaction_id,
-                "wait_for_effect": wait_for_effect,
-                "effect_timeout": effect_timeout,
-            },
-            project_root=project_path,
-            fn=live.control,
-            args=(project_path, action),
-            kwargs={
-                "interaction_id": interaction_id or None,
-                "wait_for_effect": wait_for_effect,
-                "effect_timeout": effect_timeout,
-            },
-        )
-
-    @_register_tool
-    def renforge_send_input(
-        project_path: str,
-        text: str | None = None,
-        key: str | None = None,
-        scroll: dict[str, Any] | None = None,
-        submit: bool = False,
-    ) -> dict:
-        """Send exactly one input mode: text, named key, or scroll object.
-
-        ``text`` posts character-by-character TEXTINPUT events to a focused
-        Ren'Py Input; ``submit`` optionally presses Enter after the text.
-        ``key`` accepts readable names such as enter, esc, arrows, pageup,
-        pagedown, backspace, delete, home, end, space, tab, and function keys.
-        ``scroll`` is ``{"x": ..., "y": ..., "direction": "up"|"down"}``
-        in logical game coordinates, with optional integer ``amount``.
-        Exactly one of text, key, and scroll must be supplied.
-        """
-        return _log_tool_call(
-            name="renforge_send_input",
-            params={
-                "project_path": project_path,
-                "text": text,
-                "key": key,
-                "scroll": scroll,
-                "submit": submit,
-            },
-            project_root=project_path,
-            fn=live.send_input,
-            args=(project_path,),
-            kwargs={
-                "text": text,
-                "key": key,
-                "scroll": scroll,
-                "submit": submit,
-            },
-        )
-
-    @_register_tool
-    def renforge_saves(
-        project_path: str,
-        action: str,
-        slot: str | None = None,
-        extra_info: str | None = None,
-        regexp: str | None = None,
-    ) -> dict:
-        """Save, load, or list named save slots without screenshot payloads."""
-        return _log_tool_call(
-            name="renforge_saves",
-            params={
-                "project_path": project_path,
-                "action": action,
-                "slot": slot,
-                "extra_info": extra_info,
-                "regexp": regexp,
-            },
-            project_root=project_path,
-            fn=live.saves,
-            args=(project_path, action),
-            kwargs={"slot": slot, "extra_info": extra_info, "regexp": regexp},
-        )
-
-    @_register_tool
-    def renforge_list_choices(project_path: str) -> dict:
-        """List the on-screen menu choices (text + index)."""
-        return _log_tool_call(
-            name="renforge_list_choices",
-            params={"project_path": project_path},
-            project_root=project_path,
-            fn=live.list_choices,
-            args=(project_path,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_select_choice(
-        project_path: str,
-        text: str | None = None,
-        index: int | None = None,
-    ) -> dict:
-        """Select a menu choice by visible text (preferred) or by index."""
-        return _log_tool_call(
-            name="renforge_select_choice",
-            params={"project_path": project_path, "text": text, "index": index},
-            project_root=project_path,
-            fn=live.select_choice,
-            args=(project_path,),
-            kwargs={
-                "text": text or None,
-                "index": index if isinstance(index, int) and index >= 0 else None,
-            },
-        )
-
-    @_register_tool
-    def renforge_list_ui_elements(
-        project_path: str,
-        screen: str = "",
-        text: str = "",
-        element_type: str = "",
-    ) -> dict:
-        """List visible focusable Ren'Py controls with bounds and frame guard."""
-        return _log_tool_call(
-            name="renforge_list_ui_elements",
-            params={
-                "project_path": project_path,
-                "screen": screen,
-                "text": text,
-                "element_type": element_type,
-            },
-            project_root=project_path,
-            fn=live.list_ui_elements,
-            args=(project_path,),
-            kwargs={
-                "screen": screen or None,
-                "text": text or None,
-                "element_type": element_type or None,
-            },
-        )
-
-    @_register_tool
-    def renforge_click_element(
-        project_path: str,
-        text: str = "",
-        element_id: str = "",
-        screen: str = "",
-        exact: bool = False,
-        expected_frame_id: str = "",
-        interaction_id: str = "",
-        wait_for_effect: bool = False,
-        effect_timeout: float = 5.0,
-    ) -> dict:
-        """Click a visible control by text/id, guarded against a stale frame.
-
-        Returns received_by when another control owns the hit point. With
-        wait_for_effect=true, wait for a correlated business event such as
-        quick_save.completed when the clicked action is recognizable.
-        """
-        return _log_tool_call(
-            name="renforge_click_element",
-            params={
-                "project_path": project_path,
-                "text": text,
-                "element_id": element_id,
-                "screen": screen,
-                "exact": exact,
-                "expected_frame_id": expected_frame_id,
-                "interaction_id": interaction_id,
-                "wait_for_effect": wait_for_effect,
-                "effect_timeout": effect_timeout,
-            },
-            project_root=project_path,
-            fn=live.click_element,
-            args=(project_path,),
-            kwargs={
-                "text": text or None,
-                "element_id": element_id or None,
-                "screen": screen or None,
-                "exact": exact,
-                "expected_frame_id": expected_frame_id or None,
-                "interaction_id": interaction_id or None,
-                "wait_for_effect": wait_for_effect,
-                "effect_timeout": effect_timeout,
-            },
-        )
-
-    @_register_tool
-    def renforge_hover_element(
-        project_path: str,
-        text: str = "",
-        element_id: str = "",
-        screen: str = "",
-        exact: bool = False,
-        expected_frame_id: str = "",
-    ) -> dict:
-        """Move the pointer over a visible control without clicking it."""
-        return _log_tool_call(
-            name="renforge_hover_element",
-            params={
-                "project_path": project_path,
-                "text": text,
-                "element_id": element_id,
-                "screen": screen,
-                "exact": exact,
-                "expected_frame_id": expected_frame_id,
-            },
-            project_root=project_path,
-            fn=live.hover_element,
-            args=(project_path,),
-            kwargs={
-                "text": text or None,
-                "element_id": element_id or None,
-                "screen": screen or None,
-                "exact": exact,
-                "expected_frame_id": expected_frame_id or None,
-            },
-        )
-
-    @_register_tool
-    def renforge_get_ui_element_bounds(
-        project_path: str,
-        text: str = "",
-        element_id: str = "",
-        screen: str = "",
-        exact: bool = False,
-        expected_frame_id: str = "",
-    ) -> dict:
-        """Report focus bounds and rendered painted bounds for a UI element."""
-        return _log_tool_call(
-            name="renforge_get_ui_element_bounds",
-            params={
-                "project_path": project_path,
-                "text": text,
-                "element_id": element_id,
-                "screen": screen,
-                "exact": exact,
-                "expected_frame_id": expected_frame_id,
-            },
-            project_root=project_path,
-            fn=live.get_ui_element_bounds,
-            args=(project_path,),
-            kwargs={
-                "text": text or None,
-                "element_id": element_id or None,
-                "screen": screen or None,
-                "exact": exact,
-                "expected_frame_id": expected_frame_id or None,
-            },
-        )
-
-    @_register_tool
-    def renforge_click_at(
-        project_path: str,
-        x: float,
-        y: float,
-        expected_frame_id: str = "",
-        expected_state: dict[str, Any] | None = None,
-        coordinate_space: str = "logical",
-    ) -> dict:
-        """Click screen coordinates with optional frame/state safety guards."""
-        return _log_tool_call(
-            name="renforge_click_at",
-            params={
-                "project_path": project_path,
-                "x": x,
-                "y": y,
-                "expected_frame_id": expected_frame_id,
-                "expected_state": expected_state,
-                "coordinate_space": coordinate_space,
-            },
-            project_root=project_path,
-            fn=live.click_at,
-            args=(project_path, x, y),
-            kwargs={
-                "expected_frame_id": expected_frame_id or None,
-                "expected_state": expected_state,
-                "coordinate_space": coordinate_space,
-            },
-        )
-
-    @_register_tool
-    def renforge_get_displayable_bounds(
-        project_path: str,
-        tag: str,
-        layer: str = "",
-    ) -> dict:
-        """Report where a shown image tag was rendered, in logical coordinates."""
-        return _log_tool_call(
-            name="renforge_get_displayable_bounds",
-            params={"project_path": project_path, "tag": tag, "layer": layer},
-            project_root=project_path,
-            fn=live.get_displayable_bounds,
-            args=(project_path, tag),
-            kwargs={"layer": layer or None},
-        )
-
-    @_register_tool
-    def renforge_position_element(
-        project_path: str,
-        tag: str,
-        xpos: int | float | None = None,
-        ypos: int | float | None = None,
-        xanchor: int | float | None = None,
-        yanchor: int | float | None = None,
-        xalign: int | float | None = None,
-        yalign: int | float | None = None,
-        xoffset: int | float | None = None,
-        yoffset: int | float | None = None,
-        zoom: float | None = None,
-        rotate: float | None = None,
-        layer: str = "",
-    ) -> dict:
-        """Reposition a shown image tag live and return its new logical bounds.
-
-        Provide at least one placement field. The tag keeps its current
-        attributes. Positions follow Ren'Py's rule: an integer is absolute
-        pixels (``xpos=600`` is 600px), a float is a fraction of the screen
-        (``xpos=0.5`` is the centre). Use this to converge on coordinates
-        interactively, then write the final values into the ``.rpy`` script.
-        """
-        placement = {
-            "xpos": xpos,
-            "ypos": ypos,
-            "xanchor": xanchor,
-            "yanchor": yanchor,
-            "xalign": xalign,
-            "yalign": yalign,
-            "xoffset": xoffset,
-            "yoffset": yoffset,
-            "zoom": zoom,
-            "rotate": rotate,
-        }
-        placement = {key: value for key, value in placement.items() if value is not None}
-        return _log_tool_call(
-            name="renforge_position_element",
-            params={"project_path": project_path, "tag": tag, "layer": layer, **placement},
-            project_root=project_path,
-            fn=live.position_element,
-            args=(project_path, tag),
-            kwargs={"layer": layer or None, **placement},
-        )
-
-    @_register_tool
-    def renforge_diff_screenshots(
-        project_path: str,
-        before_path: str,
-        after_path: str = "",
-        threshold: int = 0,
-    ) -> dict:
-        """Diff two frames and return the bounding box of what changed.
-
-        ``before_path`` is a saved PNG. ``after_path`` is another saved PNG, or
-        empty to diff against the current live frame. Use it to measure how far
-        an element moved or to confirm a tweak left everything else untouched.
-        """
-        def _diff() -> dict:
-            from .image_ops import diff_images
-
-            before = Path(before_path).expanduser()
-            if not before.is_absolute():
-                before = Path(project_path).expanduser() / before
-            if after_path:
-                after: Any = Path(after_path).expanduser()
-                if not after.is_absolute():
-                    after = Path(project_path).expanduser() / after
-            else:
-                try:
-                    after = live.screenshot_png(project_path)
-                except FileNotFoundError:
-                    return {"ok": False, "error": "no running game; call renforge_launch first"}
-            return diff_images(before, after, threshold=threshold)
-
-        return _log_tool_call(
-            name="renforge_diff_screenshots",
-            params={
-                "project_path": project_path,
-                "before_path": before_path,
-                "after_path": after_path,
-                "threshold": threshold,
-            },
-            project_root=project_path,
-            fn=_diff,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_eval(project_path: str, expr: str) -> dict:
-        """Evaluate a Python expression in the running game's store namespace."""
-        return _log_tool_call(
-            name="renforge_eval",
-            params={"project_path": project_path, "expr": expr},
-            project_root=project_path,
-            fn=live.eval_expr,
-            args=(project_path, expr),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_set_var(project_path: str, name: str, value: Any) -> dict:
-        """Set a variable in the running game's store namespace."""
-        return _log_tool_call(
-            name="renforge_set_var",
-            params={"project_path": project_path, "name": name, "value": value},
-            project_root=project_path,
-            fn=live.set_var,
-            args=(project_path, name, value),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_get_var(project_path: str, name: str) -> dict:
-        """Read a variable from the running game's store."""
-        return _log_tool_call(
-            name="renforge_get_var",
-            params={"project_path": project_path, "name": name},
-            project_root=project_path,
-            fn=live.get_var,
-            args=(project_path, name),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_poll_events(project_path: str, since: int = 0) -> dict:
-        """Return pushed events (dialogue, labels, exceptions) newer than `since`."""
-        return _log_tool_call(
-            name="renforge_poll_events",
-            params={"project_path": project_path, "since": since},
-            project_root=project_path,
-            fn=live.poll_events,
-            args=(project_path,),
-            kwargs={"since": since},
-        )
-
-    @_register_tool
-    def renforge_get_errors(project_path: str, since: int = 0) -> dict:
-        """Return recent bridge exceptions or bounded crash-file diagnostics."""
-        return _log_tool_call(
-            name="renforge_get_errors",
-            params={"project_path": project_path, "since": since},
-            project_root=project_path,
-            fn=live.get_errors,
-            args=(project_path,),
-            kwargs={"since": since},
-        )
-
-    @_register_tool
-    def renforge_wait_until(
-        project_path: str,
-        label: str | None = None,
-        screen: str | None = None,
-        expr: str | None = None,
-        timeout: float = 30.0,
-        interval: float = 0.5,
-        state_profile: str = "interaction",
-        include: list[str] | None = None,
-        max_depth: int = 3,
-        max_items: int = 50,
-        max_output_bytes: int = 8192,
-    ) -> dict:
-        """Wait for exactly one label, screen, or expression condition.
-
-        Returns a compact state by default (state_profile='interaction').
-        Pass include for extra fields/variables; use state_profile='full'
-        only when the complete store is required.
-        """
-        def _wait() -> dict:
-            if isinstance(timeout, bool) or not isinstance(timeout, (int, float)):
-                return {"ok": False, "error": "timeout must be a finite non-negative number"}
-            if not math.isfinite(float(timeout)) or timeout < 0:
-                return {"ok": False, "error": "timeout must be a finite non-negative number"}
-            if timeout > 120:
-                return {"ok": False, "error": "timeout must be <= 120 seconds"}
-            return live.wait_until(
-                project_path,
-                label=label,
-                screen=screen,
-                expr=expr,
-                timeout=timeout,
-                interval=interval,
-                state_profile=state_profile,
-                include=include,
-                max_depth=max_depth,
-                max_items=max_items,
-                max_output_bytes=max_output_bytes,
-            )
-
-        return _log_tool_call(
-            name="renforge_wait_until",
-            params={
-                "project_path": project_path,
-                "label": label,
-                "screen": screen,
-                "expr": expr,
-                "timeout": timeout,
-                "interval": interval,
-                "state_profile": state_profile,
-                "include": include,
-                "max_depth": max_depth,
-                "max_items": max_items,
-                "max_output_bytes": max_output_bytes,
-            },
-            project_root=project_path,
-            fn=_wait,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_hit_test(
-        project_path: str,
-        x: float,
-        y: float,
-        coordinate_space: str = "logical",
-    ) -> dict:
-        """Inspect the interactive focus stack at a coordinate.
-
-        Returns topmost and underneath focusable controls so agents can detect
-        transparent overlays that intercept clicks.
-        """
-        return _log_tool_call(
-            name="renforge_hit_test",
-            params={
-                "project_path": project_path,
-                "x": x,
-                "y": y,
-                "coordinate_space": coordinate_space,
-            },
-            project_root=project_path,
-            fn=live.hit_test,
-            args=(project_path, x, y),
-            kwargs={"coordinate_space": coordinate_space},
-        )
-
-    @_register_tool
-    def renforge_scene_tree(
-        project_path: str,
-        detail: str = "semantic",
-        layers: list[str] | None = None,
-        types: list[str] | None = None,
-        screen: str = "",
-        ids: list[str] | None = None,
-        include: list[str] | None = None,
-        format: str = "json",
-        save_as: str = "",
-        diff_against: str = "",
-        max_output_depth: int = 6,
-        max_items: int = 50,
-        max_output_bytes: int = 65_536,
-    ) -> dict:
-        """Perceive the whole scene as structured data (logical coordinates).
-
-        Unlike `renforge_list_ui_elements` (focusables only), this reports every
-        layer displayable, focusable control, and text block with `id`, `type`,
-        `bounds`, `center`, `zorder` and `screen`. Every reply carries an
-        `omitted` completeness hint. `detail` is `semantic` (default), `layout`
-        or `raw`; `layers`/`types`/`screen`/`ids` scope the result.
-        `include=["color","style","overflow"]` opts into composited colour,
-        declared style, and best-effort text overflow. `format="wireframe"` adds
-        an ASCII map. `save_as` persists a snapshot under `.renforge/scenes/`;
-        `diff_against` diffs the live scene against a saved snapshot.
-        """
-        return _log_tool_call(
-            name="renforge_scene_tree",
-            params={
-                "project_path": project_path,
-                "detail": detail,
-                "layers": layers,
-                "types": types,
-                "screen": screen,
-                "ids": ids,
-                "include": include,
-                "format": format,
-                "save_as": save_as,
-                "diff_against": diff_against,
-                "max_items": max_items,
-                "max_output_depth": max_output_depth,
-                "max_output_bytes": max_output_bytes,
-            },
-            project_root=project_path,
-            fn=live.scene_tree,
-            args=(project_path,),
-            kwargs={
-                "detail": detail or None,
-                "layers": layers or None,
-                "types": types or None,
-                "screen": screen or None,
-                "ids": ids or None,
-                "include": include or None,
-                "format": format or "json",
-                "save_as": save_as or None,
-                "diff_against": diff_against or None,
-                "max_items": max_items,
-                "max_output_depth": max_output_depth,
-                "max_output_bytes": max_output_bytes,
-            },
-        )
-
-    @_register_tool
-    def renforge_measure(
-        project_path: str,
-        action: str,
-        targets: list[Any],
-        within: Any = None,
-        tolerance: float | None = None,
-    ) -> dict:
-        """Measure pixel relationships between scene nodes, without eyes.
-
-        `action` is one of `align`, `gap`, `distribute`, `center`, `overlap`,
-        `fit`, `contrast`. Each target (and `within`) is a `renforge_scene_tree`
-        node `id` (string, resolved live) or a literal bounds object
-        `{x,y,width,height}`. Returns actionable deltas in logical pixels; when
-        `tolerance` is given, adds a `pass` verdict. `contrast` samples the live
-        frame and reports a WCAG ratio (one target = its internal fg/bg, two
-        targets = between the two elements).
-        """
-        return _log_tool_call(
-            name="renforge_measure",
-            params={
-                "project_path": project_path,
-                "action": action,
-                "targets": targets,
-                "within": within,
-                "tolerance": tolerance,
-            },
-            project_root=project_path,
-            fn=live.measure,
-            args=(project_path,),
-            kwargs={
-                "action": action,
-                "targets": targets,
-                "within": within,
-                "tolerance": tolerance,
-            },
-        )
-
-    @_register_tool
-    def renforge_run_scenario(
-        project_path: str,
-        steps: list[dict[str, Any]],
-        name: str = "scenario",
-        timeout: float = 30.0,
-        stop_on_failure: bool = True,
-        state_profile: str = "minimal",
-        capture_on_failure: bool = True,
-    ) -> dict:
-        """Run a multi-step live scenario (click/wait/assert/...) in one call.
-
-        On failure, captures a screenshot and compact diagnostics automatically.
-        Supported step actions: set, eval, click, click_at, advance, scroll,
-        wait, assert, select_choice, capture, save, load, control, send_input.
-        """
-        return _log_tool_call(
-            name="renforge_run_scenario",
-            params={
-                "project_path": project_path,
-                "name": name,
-                "timeout": timeout,
-                "stop_on_failure": stop_on_failure,
-                "state_profile": state_profile,
-                "capture_on_failure": capture_on_failure,
-                "steps": steps,
-            },
-            project_root=project_path,
-            fn=live.run_scenario,
-            args=(project_path,),
-            kwargs={
-                "steps": steps,
-                "name": name,
-                "timeout": timeout,
-                "stop_on_failure": stop_on_failure,
-                "state_profile": state_profile,
-                "capture_on_failure": capture_on_failure,
-            },
-        )
-
-    @_register_tool
-    def renforge_autopilot(project_path: str, max_runs: int = 16, max_steps: int = 60) -> dict:
-        """Auto-play the game across all branches; report label coverage and crashes."""
-        return _log_tool_call(
-            name="renforge_autopilot",
-            params={"project_path": project_path, "max_runs": max_runs, "max_steps": max_steps},
-            project_root=project_path,
-            fn=live.run_autopilot,
-            args=(project_path,),
-            kwargs={"max_runs": max_runs, "max_steps": max_steps},
-        )
-
-    # --- assets / translation / build / docs (SDK-backed, static) ---
-
-    @_register_tool
-    def renforge_assets(project_path: str) -> dict:
-        """Find orphaned and missing image/audio assets in the project."""
-        return _log_tool_call(
-            name="renforge_assets",
-            params={"project_path": project_path},
-            project_root=project_path,
-            fn=project_ops.assets,
-            args=(project_path,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_languages(project_path: str) -> dict:
-        """List translation languages present under game/tl/."""
-        return _log_tool_call(
-            name="renforge_languages",
-            params={"project_path": project_path},
-            project_root=project_path,
-            fn=project_ops.languages,
-            args=(project_path,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_translation_stats(project_path: str, language: str) -> dict:
-        """Report missing dialogue/string translation counts for a language."""
-        return _log_tool_call(
-            name="renforge_translation_stats",
-            params={"project_path": project_path, "language": language},
-            project_root=project_path,
-            fn=project_ops.translation_stats,
-            args=(project_path, language),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_generate_translations(project_path: str, language: str) -> dict:
-        """Generate/update translation files for a language (writes game/tl/<language>/)."""
-        return _log_tool_call(
-            name="renforge_generate_translations",
-            params={"project_path": project_path, "language": language},
-            project_root=project_path,
-            fn=project_ops.generate_translations,
-            args=(project_path, language),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_export_dialogue(project_path: str, language: str = "None") -> dict:
-        """Export the game's dialogue as plain text."""
-        return _log_tool_call(
-            name="renforge_export_dialogue",
-            params={"project_path": project_path, "language": language},
-            project_root=project_path,
-            fn=project_ops.export_dialogue,
-            args=(project_path, language),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_web_build(project_path: str, destination: str = "") -> dict:
-        """Package the project as a browser-playable build (needs the web DLC)."""
-        return _log_tool_call(
-            name="renforge_web_build",
-            params={"project_path": project_path, "destination": destination},
-            project_root=project_path,
-            fn=project_ops.web_build,
-            args=(project_path,),
-            kwargs={"destination": destination},
-        )
-
-    @_register_tool
-    def renforge_distribute(project_path: str, package: str = "", destination: str = "") -> dict:
-        """Build desktop distributions (e.g. package='pc', 'mac', 'linux')."""
-        return _log_tool_call(
-            name="renforge_distribute",
-            params={"project_path": project_path, "package": package, "destination": destination},
-            project_root=project_path,
-            fn=project_ops.distribute,
-            args=(project_path,),
-            kwargs={"package": package, "destination": destination},
-        )
-
-    @_register_tool
-    def renforge_search_docs(query: str) -> dict:
-        """Search Ren'Py's offline documentation for a keyword."""
-        return _log_tool_call(
-            name="renforge_search_docs",
-            params={"query": query},
-            project_root=None,
-            fn=project_ops.search_docs,
-            args=(query,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_get_doc(topic: str) -> dict:
-        """Read a Ren'Py documentation page as plain text (e.g. topic='cli')."""
-        return _log_tool_call(
-            name="renforge_get_doc",
-            params={"topic": topic},
-            project_root=None,
-            fn=project_ops.get_doc,
-            args=(topic,),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_list_docs() -> dict:
-        """List available Ren'Py documentation topics."""
-        return _log_tool_call(
-            name="renforge_list_docs",
-            params={},
-            project_root=None,
-            fn=project_ops.list_docs,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_screenshot(
-        project_path: str,
-        width: int = 0,
-        height: int = 0,
-        crop_x: int = 0,
-        crop_y: int = 0,
-        crop_width: int = 0,
-        crop_height: int = 0,
-        scale: float = 1.0,
-        grid: int = 0,
-        crosshair_x: int = -1,
-        crosshair_y: int = -1,
-        rulers: bool = False,
-    ):
-        """Capture a game frame, optionally resizing, cropping, and zooming it.
-
-        Measurement guides help pixel-perfect placement: ``grid`` draws lines
-        every N pixels, ``rulers`` labels those steps along the edges, and
-        ``crosshair_x``/``crosshair_y`` mark a point. Capture at the game's
-        logical resolution (``width``/``height``) so the labels read as logical
-        coordinates. Passing only one of ``width``/``height`` keeps the game's
-        aspect ratio.
-        """
-        def _tool() -> Any:
-            try:
-                if width < 0 or height < 0:
-                    raise ValueError("width and height must be non-negative")
-                if (crosshair_x < 0) != (crosshair_y < 0):
-                    raise ValueError("crosshair_x and crosshair_y must be provided together")
-                if width or height:
-                    png = live.screenshot_png(project_path, width=width, height=height)
-                else:
-                    png = live.screenshot_png(project_path)
-                if crop_width or crop_height or crop_x or crop_y or scale != 1.0:
-                    from .image_ops import transform_png
-
-                    png = transform_png(
-                        png,
-                        crop_x=crop_x,
-                        crop_y=crop_y,
-                        crop_width=crop_width,
-                        crop_height=crop_height,
-                        scale=scale,
-                    )
-                if grid or rulers or crosshair_x >= 0:
-                    from .image_ops import annotate_png
-
-                    png = annotate_png(
-                        png,
-                        grid=grid,
-                        rulers=rulers,
-                        crosshair=(crosshair_x, crosshair_y) if crosshair_x >= 0 else None,
-                    )
-            except FileNotFoundError:
-                return {"ok": False, "error": "no running game; call renforge_launch first"}
-            except Exception as exc:  # pragma: no cover - defensive
-                return {"ok": False, "error": f"{type(exc).__name__}: {exc}"}
-
-            # Return a raw MCP content block: helper classes like fastmcp.Image
-            # moved between fastmcp versions, and an Image object from the
-            # wrong package gets stringified instead of rendered.
-            return _png_content(png)
-
-        return _log_tool_call(
-            name="renforge_screenshot",
-            params={
-                "project_path": project_path,
-                "width": width,
-                "height": height,
-                "crop_x": crop_x,
-                "crop_y": crop_y,
-                "crop_width": crop_width,
-                "crop_height": crop_height,
-                "scale": scale,
-                "grid": grid,
-                "crosshair_x": crosshair_x,
-                "crosshair_y": crosshair_y,
-                "rulers": rulers,
-            },
-            project_root=project_path,
-            fn=_tool,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_capture_screenshot(
-        project_path: str,
-        name: str = "capture",
-        width: int = 0,
-        height: int = 0,
-        crop_x: int = 0,
-        crop_y: int = 0,
-        crop_width: int = 0,
-        crop_height: int = 0,
-        scale: float = 1.0,
-        grid: int = 0,
-        crosshair_x: int = -1,
-        crosshair_y: int = -1,
-        rulers: bool = False,
-    ) -> dict:
-        """Persist a screenshot under the project's controlled capture directory."""
-        def _capture() -> dict:
-            from io import BytesIO
-            from PIL import Image
-            from .captures import validate_capture_name, write_project_capture
-
-            validate_capture_name(name)
-            if width < 0 or height < 0:
-                raise ValueError("width and height must be non-negative")
-            if (crosshair_x < 0) != (crosshair_y < 0):
-                raise ValueError("crosshair_x and crosshair_y must be provided together")
-            png = live.screenshot_png(project_path, width=width, height=height)
-            if crop_width or crop_height or crop_x or crop_y or scale != 1.0:
-                from .image_ops import transform_png
-                png = transform_png(png, crop_x=crop_x, crop_y=crop_y,
-                                    crop_width=crop_width, crop_height=crop_height,
-                                    scale=scale)
-            if grid or rulers or crosshair_x >= 0:
-                from .image_ops import annotate_png
-                png = annotate_png(png, grid=grid, rulers=rulers,
-                                   crosshair=(crosshair_x, crosshair_y) if crosshair_x >= 0 else None)
-            project_root, target = write_project_capture(project_path, name, png)
-            with Image.open(BytesIO(png)) as image:
-                size = {"width": image.width, "height": image.height}
-            return {"ok": True, "name": name, "path": str(target),
-                    "relative_path": str(target.relative_to(project_root)),
-                    "sha256": hashlib.sha256(png).hexdigest(), "size": size}
-
-        return _log_tool_call(
-            name="renforge_capture_screenshot",
-            params={"project_path": project_path, "name": name, "width": width, "height": height},
-            project_root=project_path, fn=_capture, args=(), kwargs={})
-
-    @_register_tool
-    def renforge_estimate_translation(
-        before_path: str,
-        after_path: str,
-        region_x: int = 0,
-        region_y: int = 0,
-        region_width: int = 0,
-        region_height: int = 0,
-        threshold: int = 16,
-        max_shift: int = 64,
-    ) -> dict:
-        """Estimate stable visual translation between two saved frames."""
-        def _estimate() -> dict:
-            from .image_ops import estimate_translation
-            region = None
-            if region_width or region_height or region_x or region_y:
-                region = (region_x, region_y, region_width, region_height)
-            return estimate_translation(
-                before_path,
-                after_path,
-                region=region,
-                threshold=threshold,
-                max_shift=max_shift,
-            )
-
-        return _log_tool_call(
-            name="renforge_estimate_translation",
-            params={"before_path": before_path, "after_path": after_path},
-            project_root=None,
-            fn=_estimate,
-            args=(),
-            kwargs={},
-        )
-
-    @_register_tool
-    def renforge_find_image_on_screen(
-        project_path: str,
-        template_path: str,
-        threshold: float = 0.92,
-        max_matches: int = 20,
-        region_x: int = 0,
-        region_y: int = 0,
-        region_width: int = 0,
-        region_height: int = 0,
-    ) -> dict:
-        """Find a template image in the current frame and return its bounds."""
-        def _find() -> dict:
-            from .image_ops import find_image_matches
-
-            if (region_width == 0) != (region_height == 0):
-                raise ValueError("region_width and region_height must be provided together")
-            if (region_x or region_y) and not (region_width and region_height):
-                raise ValueError("region coordinates require region_width and region_height")
-            screenshot = live.screenshot_png(project_path)
-            template = Path(template_path).expanduser()
-            if not template.is_absolute():
-                template = Path(project_path).expanduser() / template
-            region = (
-                (region_x, region_y, region_width, region_height)
-                if region_width and region_height
-                else None
-            )
-            result = find_image_matches(
-                screenshot,
-                template,
-                threshold=threshold,
-                max_matches=max_matches,
-                region=region,
-            )
-            result["frame_id"] = hashlib.sha256(screenshot).hexdigest()
-            result["coordinate_space"] = "screenshot"
-            result["click_hint"] = {
-                "coordinate_space": "screenshot",
-                "expected_frame_id": result["frame_id"],
-            }
-            return result
-
-        return _log_tool_call(
-            name="renforge_find_image_on_screen",
-            params={
-                "project_path": project_path,
-                "template_path": template_path,
-                "threshold": threshold,
-                "max_matches": max_matches,
-                "region_x": region_x,
-                "region_y": region_y,
-                "region_width": region_width,
-                "region_height": region_height,
-            },
-            project_root=project_path,
-            fn=_find,
-            args=(),
-            kwargs={},
-        )
-
-    expected_tool_names = set(TOOL_DEFINITIONS)
-    registered_names = set(registered_tools)
-    if expected_tool_names != registered_names:
-        missing = sorted(expected_tool_names - registered_names)
-        extra = sorted(registered_names - expected_tool_names)
-        raise RuntimeError(
-            f"Tool registration drift: missing={missing} extra={extra}"
-        )
+    register_all_tools(app)
 
 
 def create_app() -> Any:
